Refactorings, cleanups and comments
