+++ core/src/main/kotlin/ecs/components/AudioChannels.kt
@@ -1,0 +1,7 @@
+package ecs.components
+
+object AudioChannels {
+    const val enemyDeath = "ENEMYDEATH"
+    const val default = "DEFAULT"
+    const val simultaneous = "SIMULTANEOUS"
+}
