+++ core/src/main/kotlin/ecs/components/AudioComponent.kt
@@ -3,12 +3,6 @@ package ecs.components
 import audio.TurboSound
 import ecs.components.ai.CoolDownComponent
 
-object AudioChannels {
-    const val enemyDeath = "ENEMYDEATH"
-    const val default = "DEFAULT"
-    const val simultaneous = "SIMULTANEOUS"
-}
-
 class AudioComponent: CoolDownComponent() {
     var takeDistanceIntoAccount = true
     var channel = AudioChannels.simultaneous
