+++ core/src/main/kotlin/ecs/components/VehicleComponents.kt
@@ -2,6 +2,7 @@ package ecs.components
 
 import com.badlogic.ashley.core.Component
 
+//TODO: Change these to Intents!
 class EnterVehicleComponent: Component
 class LeaveVehicleComponent: Component
 class IsInVehicleComponent: Component
