+++ core/src/main/kotlin/ecs/components/intent/IntendsTo.kt
@@ -1,0 +1,7 @@
+package ecs.components.intent
+
+sealed class IntendsTo {
+    object DoNothing: IntendsTo()
+    object ToggleBuildMode : IntendsTo()
+    object Build : IntendsTo()
+}
